更新2020.3.8 21:18

+++ Demo-P1/Toy.cpp
@@ -19,7 +19,7 @@ string Toy::getName() const{
 	return name;
 }
 
-int Toy::getPrice() const{
+float Toy::getPrice() const{
 	return price * discount;
 }
